+++ tests/support/target_script.py
@@ -4,7 +4,9 @@ target_script
 
 The target for testing.
 """
-from ec.ec import task, arg, group, call
+from ec.ec import task, arg, group, call, module
+
+# module(name='target_script')
 
 @task(alias='t1')
 @arg(type=int, desc= 'Value for arg1')